backend/requirements.txt
APScheduler==3.11.0
blinker==1.9.0
boto3==1.39.14
click==8.1.3
Flask==2.3.2
gunicorn==20.1.0
python-dotenv==1.0.0
pymongo==4.5.0
motor==3.1.2